fix delta query
